vmgr: fix deadlock on flocking data.img
if you open the orbstack data.img file in finder, a file lock will be held by `_diskimagesiod` (tested on MacOS 15.3 Beta 24D5034f; not present on 15.3.1 24D70)

ensureDataLock() will spawn a goroutine to try to flock data.img, then return an error if a 10s timeout is exceeded. on return, it'll try to close the fd, which will wait on the routine running flock, resulting in a deadlock.

to fix, fatal error and exit(1) if data.img lock exceeds deadline