scon: cert injector: only add missing certs to bundle
if the the host has new certs constantly being added to the keychain
(looking at you, Avast), then the CA certs bundle in the container will
grow by O(N^2) if we re-add the entire set of PEMs to the bundle each
time the set changes.

also, GetExtraCaCertificates always includes the orb.local root, so no
need to add it separately.