scon: move TLS certificate generation to vmgr for security
since the OrbStack root CA is installed and trusted for all SSL purposes
on the system (at the user level), it's extremely powerful; wrongful
possession of it allows MITMing all TLS traffic on the host when combined
with a proxy, VPN, or some other way to intercept DNS or network traffic.

currently, scon has the private key in RAM and enforces a *.local
restriction on generated certificates. a machine escape would compromise
the private key.

we usually consider OrbStack machines to be trusted; if they're privileged,
they should be able to do anything the user can on macOS.

however, accessing the OrbStack CA private key is not something that any
typical user process can do on macOS. it's tied to our keychain access
group, which requires a restricted entitlement and provisioning profile.
so only legitimately OrbStack-signed code can access it.

by using vmgr as a "sandbox" / enclave and moving all private key usage
to the host, this becomes far more secure than other development CA setups:
the only possible way to bypass the *.local restriction is to find a
memory safety bug in vmgr.

in the future we could also consider making the *.local restriction
tighter by doing a DNS lookup in vmgr to make sure that it points to an
OrbStack IP, but that's more fragile. a sandbox process would also make
this more secure, but it seems unnecessary for now.

this obviously has a bit of a perf hit (1-15ms per cert), mostly because
of the keychain part:
🌲 vmgr | DEBU[04-01 21:10:59] SARU keychain                                 duration=5.322291ms
🌲 vmgr | DEBU[04-01 21:10:59] SARU parse cert                               duration="67.875µs"
🌲 vmgr | DEBU[04-01 21:10:59] SARU generate cert                            duration="194.292µs"
🌲 vmgr | DEBU[04-01 21:10:59] SARU marshal private key                      duration="4.666µs"

but for now, let's maximize security here and reduce chances of the key
material staying in memory; it'll usually be GC'd shortly after the
keychain operation. mmap(MAP_JIT) is not allowed in the vmgr process
because com.apple.security.cs.allow-jit isn't set, so this protects
against host memory read vulns, and code injection isn't possible.