scon: generate unique SSH host keys for security
- we've always used hard-coded SSH host keys because our SSH was not
  intended to be used for security (internal uses unix socket; the worst
  we have is localhost:32222)
- however, some people have been forwarding the localhost:32222 server out
  to LAN using ssh/socat
- that's not ideal because public, hard-coded host keys means that it can
  be MITMed

but regenerating keys for everyone is bad UX because it breaks known_hosts,
and patching known_hosts doesn't help because other clients like Jetbrains
save it elsewhere.

instead, keep the old hard-coded keys for existing users, and only
generate unique keys for new installs. if/when we add an official setting
to make the SSH server public, the flag that gets set on migration allows
us to detect this case and regenerate keys when the setting is enabled.