vmgr: add 60-second pause-on-sleep debounce to mitigate IOKit hangs
buggy apps that aren't responding to IOKit WillSleep events can cause all
event delivery to hang for up to 30s, which can cause the user to see
frozen containers for up to 30s if they close the lid and then almost
immediately open it, as we will have gotten the WillSleep event but won't
get the WillPowerOn event for 30 more seconds.

to mitigate this, add a debounce to only pause on sleep if 60 seconds of
monotonic (awake) time have passed since the WillSleep event. this still
helps with battery drain because we'll pause the containers as soon as
we hit 60 seconds of awake time, which we'll reach pretty quickly if the
machine is waking up often enough to cause significant drain.