scon: don't report that machine is running until creation is finished
this solves two problems:

a. it's common to create a machine in the gui, wait for it to complete, and then ssh in. if a machine isn't finished initializing and a user tries to connect too early, login will fail.

b. it's possible to start creating a machine, persist its state as running before initialization completes, and then exit scon. on the next start, the machine will be stuck in an uninitialized state.