virtue: support resetting virtio devices in active state
this is not an abnormal action; resetting is one way to put the devices
into init state. deactivating them first is not required.

kernel 6.14.9+ resets all devices on poweroff (commit aee42f3d57bf). this
is fine for shutdown time because it just changes when we do the resets
(from VmmExitObserver to kernel poweroff)