virtue: remove cc 1.1.31 patch
it's not doing anything and I have no problems without it.

warning: Patch `cc v1.1.31 (https://github.com/rust-lang/cc-rs.git?tag=cc-v1.1.31#c07ab0ea)` was not used in the crate graph.
Check that the patched package version and available features are compatible
with the dependency requirements. If the patch has a different version from
what is locked in the Cargo.lock file, run `cargo update` to use the new
version. This may also occur with an optional dependency that is not enabled.