scon: rewrite container locking
scon's container locking is a disaster and it's unclear what c.mu is
even supposed to protect.

well, it's still kind of unclear. but what is at least more clear now is
that all the *runtime* state is atomically published as a set of read-only
client objects. as a result, locking, bpf, etc. no longer need locking,
which simplifies things greatly.

also get rid of the janky SIGPWR warning signal thing for docker agent
stops.